Removed kmp4free since it has its own repo now here: https://github.com/handstandsam/kmp4free
Updated dependency Guard baseline.

Progress
