+++ .github/workflows/build_and_test.yaml
@@ -52,10 +52,10 @@ jobs:
         run: ./gradlew dependencyGuard --scan --stacktrace
 
       - name: Run All Checks - KMP4FREE Disabled
-        run: ./gradlew check -Pkmp4free.multiplatform=false -Pkmp4free.ios=false -Pkmp4free.js=false --scan --stacktrace
+        run: ./gradlew check -Pkmp4free=false -Pkmp4free.ios=false -Pkmp4free.js=false --scan --stacktrace
 
       - name: Run All Checks with Multiplatform - KMP4FREE Enabled
-        run: ./gradlew check -Pkmp4free.multiplatform=true -Pkmp4free.ios=true -Pkmp4free.js=true --scan --stacktrace
+        run: ./gradlew check -Pkmp4free=true -Pkmp4free.ios=true -Pkmp4free.js=true --scan --stacktrace
 
       - name: (Fail-only) Bundle test reports
         if: failure()
