+++ convention-plugins/src/main/kotlin/com/handstandsam/convention/plugins/JvmLibConventionPlugin.kt
@@ -1,9 +1,8 @@
 package com.handstandsam.convention.plugins
 
-import org.gradle.api.JavaVersion
+import com.handstandsam.convention.plugins.Kmp4FreeLibConventionPlugin.Companion.setJvmVersionCompatibility
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPluginExtension
 
 /**
  * Standard Kotlin JVM Plugin
@@ -12,14 +11,5 @@ class JvmLibConventionPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         target.plugins.apply("org.jetbrains.kotlin.jvm")
         setJvmVersionCompatibility(target)
-    }
-
-    companion object {
-        fun setJvmVersionCompatibility(target: Project) {
-            target.extensions.getByType(JavaPluginExtension::class.java).apply {
-                sourceCompatibility = JavaVersion.VERSION_1_8
-                targetCompatibility = JavaVersion.VERSION_1_8
-            }
-        }
     }
 }
