+++ gradle.properties
@@ -27,6 +27,6 @@ kotlin.mpp.stability.nowarn=true
 kotlin.mpp.enableCInteropCommonization=true
 
 # KMP4FREE
-kmp4free.multiplatform=true
+kmp4free=true
 #kmp4free.ios=true
 kmp4free.js=true
