--- kmp4free/gradle/libs.versions.toml
@@ -1,1 +1,0 @@
-../../gradle/libs.versions.toml
